ZOOKEEPER-4736: Fix nio socket fd leak if network service is down
`Socket::close` could only mark the object as closed but leave
underlying socket fd open if the socket is registered to nio selector.
So, we have to call `Selector::selectNow` in `cleanup` to close
underlying socket fd.

This could happen if network service get shutdown in socket connecting,
say, "ifdown eth0" or "service network stop" and so on.

See:
* https://github.com/openjdk/jdk/blob/jdk8-b120/jdk/src/share/classes/sun/nio/ch/SocketChannelImpl.java#L841
